xen/domain: introduce generic weak function for domain struct allocation
x86 has specific requirements about the allocation of the domain structure,
but that's not the case for ARM or likely other architectures.

Introduce a generic weak function that can be overwritten with an
architecture specific implementation if required.

No functional change intended.

For RISC-V this requires including asm/page.h in asm/mm.h, like other
arches do, so clear_page() is available in domain.c.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>