xen: declare paddr_bits on the same header on all architectures
On ARM/PowerPC/RISCV move the definition of paddr_bits from p2m.h to
processor.h.  This is a more suitable place rather than p2m.h, since the
width of physical addresses is not explicitly related to the usage of p2m.

No functional change intended.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
We could also move it to page-bits.h, but on x86 at least processor.h seems
like a best option, since the physical address width is a property of the
processor.