xen/livepatch: fixup relocations to replaced symbols
In a livepatch payload relocations will refer to included functions.  If
that function happens to be a replacement for an existing Xen function, the
relocations on the livepatch payload will use the newly introduced symbol,
rather than the old one.  This is usually fine, but if the result of the
relocation is stored for later use (for example in the domain struct),
usages of this address will lead to a page-fault once the livepatch is
reverted.

Implement a second pass over relocations once the list of replaced
functions has been loaded, and fixup any references to replaced functions
to use the old symbol address instead of the new one.  There some
sections that must be special cases to continue using the new symbol
address, as those instances must reference the newly added livepatch
content (for example the possible alternative patch sites).

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>